Move EndpointKey and ParameterKey to commonMain
